requirements.txt
Django>=4.2
daphne>=4.0
channels>=4.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-filter>=23.2
django-cors-headers>=4.0
gunicorn>=20.1
uvicorn>=0.22
pillow>=9.5

# Excel import
pandas>=2.0.0
openpyxl>=3.1.0

# AI Image Generation
requests>=2.31.0
